+++ scripts/git-subtree/wakatime/add.ps1
@@ -1,0 +1,5 @@
+git subtree add `
+    -P workspace/plugins/wakatime `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-wakatime.git `
+    dev `
+    --squash