+++ scripts/git-subtree/wakatime/push.ps1
@@ -1,0 +1,4 @@
+git subtree push `
+    -P workspace/plugins/wakatime `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-wakatime.git `
+    dev